[refact] Extending function to fit the songlist feature
mpc#stractSongFromString : Fixed a typo on item variable, it is items now.
mpc#formatSonglist: New function based on mpc#formatPlaylist.
mpc#formatSong: This function was rewritten so it can accept a parameter to overwrite its defaults
mpc#songToArray: This function was extended so the order of the fields can be passed to it

Most of the functions were refatored so they only access data from arrays and
hashes using the get function.